Add helper script to prepare `reuse` runs.
reuse is quite picky about files lingering around in a development
directory, although many of them do not belong to the project and
therefore should not be 'blessed' with a copyright or licence.

We therefore run `reuse` by `tox` and in an isolated environment, where
only the sdist package is extracted. Extracting the sdist package
however is not done by `tox`, so we use this little helper script.